feat: create counter features with progress bar and relative titles
