--- src/components/ProgressBarContainer/NotesProgressBar.tsx
@@ -1,10 +1,0 @@
-import * as S from "./styles";
-
-export function NotesProgressBar() {
-  return (
-    <S.ProgressBarContainer>
-      <S.BackgroundDefaultProgressBar />
-      <S.BackgroundPercentProgressBar />
-    </S.ProgressBarContainer>
-  );
-}
